pyproject.toml
[tool.poetry]
name = "quizzy"
version = "0.1.0"
description = "Quiz platform with Django, Channels, DRF"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "accounts" }, { include = "quizzes" }, { include = "realtime" }, { include = "submissions" }, { include = "core" }, { include = "api" }]

[tool.poetry.dependencies]
python = "^3.13"
django = "5.0.7"
djangorestframework = "3.15.2"
channels = "4.1.0"
channels-redis = "4.2.0"
redis = "5.0.7"
daphne = "4.1.0"
django-prometheus = "2.3.1"
psycopg-binary = "^3.2.9"
boto3 = "1.34.148"
django-storages = "1.14.2"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
isort = "5.13.2"
flake8 = "7.0.0"
pytest = "8.2.2"
pytest-django = "4.8.0"
coverage = "7.5.1"
mypy = "1.10.0"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
exclude = ["migrations", ".venv", "__pycache__"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "quiz_project.settings"
python_files = ["tests.py", "test_*.py"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
